% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\docType{data}
\name{area}
\alias{area}
\title{This Is A One-line Title Describing area data}
\format{x is a data frame with 50 rows and 3 columns:
\describe{
  \item{state}{names of country}
  \item{land}{acre of land}
  \item{farm}{acre of farm}
}}
\source{
These data are from my teacher Yongsu Li
(\url{http://www.stat.wisc.edu/~jgillett/327-3/2/farmLandArea.csv}) 
at UW-Madison in 2018 Nov.
}
\usage{
area
}
\description{
This is the Description paragraph. x is my favorite data set.  Feels good~
}
\keyword{datasets}
