% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_lad.R
\name{coef.las}
\alias{coef.las}
\title{Extract \deqn{coefficient of LASSO regression}}
\usage{
\method{coef}{las}(obj, ...)
}
\arguments{
\item{obj}{the object you get from lasso function}

\item{...}{further arguments passed to or from other methods}
}
\description{
Calculate \deqn{LASSO regression}
}
