% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lasso}
\alias{lasso}
\title{Calculate \deqn{LASSO regression}}
\usage{
lasso(x, y)
}
\arguments{
\item{x}{the data vector, explanatory variable}

\item{y}{the data vector, response variable}
}
\value{
\deqn{beta0,beta1}
}
\description{
Calculate \deqn{LASSO regression}
}
\details{
\code{lad} calculates
            \deqn{the parameter of LASSO regression}.
}
\examples{
obj = lasso(area$land,area$farm)
print(obj)
plot(area$land,area$farm)
abline(reg = lm(farm~land,data = area),col = "red")
abline(reg = obj,col = "navy")
legend("bottomright",legend = c("lm()","lasso()"),col = c("red","navy"))
x_new = quantile(area$land,probs = seq(0,1,1/4))
y_new = predict(obj = obj,x_new = x_new)
points(x = x_new,y = y_new, col = "green",pch = 19)
}
