% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_lad.R
\name{predict.las}
\alias{predict.las}
\title{Predict new response from lasso model \deqn{y_new_hat}}
\usage{
\method{predict}{las}(obj, x_new, ...)
}
\arguments{
\item{obj}{the object you get from lasso function}

\item{...}{further arguments passed to or from other methods}
}
\description{
Predict \deqn{y_new_hat LASSO model}
}
