% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_lad.R
\name{print.las}
\alias{print.las}
\title{Print \deqn{coefficient of LASSO regression}}
\usage{
\method{print}{las}(obj, ...)
}
\arguments{
\item{obj}{the object you get from lasso function}

\item{...}{further arguments passed to or from other methods}
}
\description{
Calculate \deqn{LASSO regression}
}
